Added mcphost container
mcphost is a wrapper around a model to detect and make MCP calls,
similar to how lightspeed-core would do it, except it comes with a
simple TUI for developers to use. We can use it to play around with our
system prompt / MCP / model combinations, without having to go through
lightspeed-core or the assisted UI.

This commit adds a new container to the assisted-chat-pod, which runs
mcphost and configures it to use the assisted-service MCP server and
Gemini, with the API keys and everything.

Use `make mcphost` to attach to the mcphost container and interact with
it. Note that if you attach to late, you might get a weird broken TUI
that if you try to use it would crash mcphost. This is actually a good
thing, because then it will automatically restart, at this point try to
attach again (this time early enough) and wait until it starts and the
TUI should be good to go.