Add initial eval tests
This commit replaces the instructions and placeholder tests with actual
eval tests that run and pass.

Currently they just do a substring test on the output of the model, and
only handle read queries (rather than creating a cluster), but this is
just a start to get something that can be improved over time.

Additionally the tests themselves currently pass if _any_ of the listed
strings are present in the output, but it would be better to get some
more flexible matching behavior (at least an AND) option for things like
the versions test.