Make OpenShift service headless
This patch changes the service created by the OpenShift template so that
it will be headless. This means that when clients resolve the name in
DNS they will get directly the IPs of the pods, and not the single IP of
the service. This is necessary because will be sending traffic from an
Envoy proxy that needs to see the IP addresses of the pods in order to
implement stateful sessions.

Related: https://issues.redhat.com/browse/MGMT-21166
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>