Use a single Containerfile for assisted-chat
We don't want to maintain two separate Containerfiles for assisted-chat.
Originally we had one for development, that uses our own built
lightspeed-stack plus installed runtime dependencies, and one for
production, that uses the official lightspeed-stack image plus installed
runtime dependencies.

For now we've decided to keep things simple and always use our own built
lightspeed-stack image, even in production.

Since we need to build with Konflux, and we don't want it to build two
separate Containerfiles, one that derives from the other, we've decided
to place a copy of the lightspeed-stack Containerfile inside our own
assisted-chat Containerfile, with a slight modification to have it not
depends on the pypi llama-stack package, but rather on the one we
maintain in our submodule.

I've also added a script that checks if the Containerfile copied from
lightspeed-stack is in sync with the original one. This is to ensure
that we don't forget to update the Containerfile in assisted-chat when
the one in lightspeed-stack changes. The script will be ran as part of
the CI pipeline, and will fail if the Containerfiles are not in sync.