Use `sub` and `preferred_username` as claims
We currently use `user_id` and `username` as claims in the
authentication configuration of the Lightspeed stack. These claims are
not available in service account tokens, which are used by our CI.

Instead we should use `sub` and `preferred_username`, which are
available in both user and service account tokens.