Use sqlite for metadata store
The metadata store remembers things like the MCP endpoint, that may
change in the llama-stack config between versions. This is annoying
because the config should take precedence. If we change the config, we
don't want it to remember the old values.

This is a llama-stack bug that needs to be fixed, and when it does this
commit needs to be reverted.

We have no value for storing this stuff in a database for now anyway, so
we move it to using a temporary sqlite database that gets thrown away
when the pod is updated with a new config.