Bump lightspeed-stack to authorization commit
Bump lightspeed-stack submodule to include the authorization changes
from https://github.com/lightspeed-core/lightspeed-stack/pull/356

Use the new authorization features to restrict access to the
assisted-chat service to, for now, Red Hat employees only.

In the future we can open this up to all authenticated users.

Also bumped to a version that no longer requires the patching of
llama-stack / lightspeed-stack to support postgres SSL, so removed the
patching steps from the Containerfile.