MGMT-21352: Update system prompt regarding static networking support
This commit updates the assisted-chat responses to correctly address user queries about static networking. Previously, the bot would confuse static networking for hosts with the manual configuration of API and Ingress VIPs.
assisted-chat new states that it doesn't support configuring static networking and instruct the user to use the UI wizard to create the cluster instead
* add eval test

Signed-off-by: Eran Cohen <eranco@redhat.com>