MGMT-21661: Tooling for Archive Data
This gives us scripts to download/extract/summarize the transcript/feedback
data that is exported to s3 from lightspeed.

It will pull all of the data about a single conversation, including feedback
and MCP tool calls into a single file per conversation. This should make it
easier to get a overall view of a conversation.