MGMT-21646: Static Networking Configuration
Allow complex static networking via the use of an MCP driven template
which serves as a base guide for generating the nmstate yaml. This can
then be amended by the model if desired so that the MCP template doesn't
need to support every single option, while still providing a general
framework within which the model works.