MGMT-21807: Relax Non-disclosure rules to enable tool calls
Occasionally the bot is making up data from tool calls without actually
doing the calls. Other times it just refuses to do the tool calls it can
already do without the user telling it explicitly to do them.

I think some of the prompt language was causing it to avoid making tool calls
in the first place because of ambiguous words.