Perform lightspeed-stack migrations
# Background

lightspeed-stack does not currently handle database migrations on its
own. This means that when we update lightspeed-stack to a new version
that requires database schema changes, we need to handle the migrations
ourselves. Otherwise lightspeed-stack will not start or will function
wrong.

This happened to us recently when we upgraded lightspeed-stack to 0.3.0
and since they added a new column, list conversation calls were failing.

# Solution

We will do the migrations using a new `migrate.py` file defined in this
repo which we will add to our containers and run right before starting
lightspeed-stack.

# Other changes

- We need to set the `LIGHTSPEED_STACK_POSTGRES_SSL_MODE` environment
  variable in the pod spec so that our migration script can connect to
  the database correctly.

- We need to set the PostgreSQL data volume mount to `:Z` so that SELinux
  does not block the database from writing to it.

- Added optional comments in the dev pod that can be uncommented to
  persist PostgreSQL data between pod restarts by using a hostPath
  volume. Also modified the `run.sh` script to map the PostgreSQL user
  inside the container to the current host user so that there are no
  permission issues when using a hostPath volume. This is conditional so
  that users who do not want to persist data do not suffer the slowdowns
  of using userns mapping.

- We need to source the `template-params.dev.env` file in our `run.sh`
  script so that the `LIGHTSPEED_STACK_POSTGRES_SSL_MODE` variable is
  available when we run `podman play kube`.