Use a script to validate static networking
Previously we used a tool validation to check that the static networking
was generated and applied. Both of these failed frequently because the
argument regex validation is very fragile.

This approach allows us full control over validating what the cluster
should look like at the end of the conversation which is all we really
care about.

Additionally the intent eval should be enough to handle the generation
tool result.