Making the name resolution part in the system prompt more strict.
Also with the current changes the name resolution part covers more cases. Previously they were thought to be "common sense".
The background of the issue is that the evaluation test
cluster_info_conv/cluster_info_tool_call failed sometimes either because
it thought that the name is the uuid, or it did not echoed the name of
the cluster when it was not able to find it.
I think echoing the name of the cluster back when not able to find it
helps the user identify typos.